TASK-46372 Rename ExoModal component in spaces administration to use a specific name
